refactored ComputerManager into ai Directory w/ three computers - HAL9000, Terminator, and Zombocom.
